Tests/LibWeb/Text/expected/ResizeObserver/disconnect.txt

Size changed: 100px x 100px
Size changed: 200px x 200px
